+++ Proyecto1_Progra4/nbproject/project.properties
@@ -12,7 +12,7 @@ build.test.classes.dir=${build.dir}/test/classes
 build.test.results.dir=${build.dir}/test/results
 build.web.dir=${build.dir}/web
 build.web.excludes=${build.classes.excludes}
-client.urlPart=/index
+client.urlPart=/Inicio
 compile.jsps=false
 conf.dir=${source.root}/conf
 debug.classpath=${build.classes.dir}:${javac.classpath}
